requirements.txt
numpy
librosa
soundfile
phonemizer
PyYAML
tqdm
torch
torchaudio
pytorch-lightning
